% Generated by build script
\name{lacity_count}
\alias{lacity_count}
\title{Get row count for a dataset query using SoQL}
\usage{
lacity_count()
}
\arguments{
\item{dataset_id}{Socrata dataset identifier}
\item{where}{Optional SoQL WHERE clause to filter}
}
\value{integer: total number of matching rows}
\description{Get row count for a dataset query using SoQL}
