% Generated by build script
\name{lacity_metadata}
\alias{lacity_metadata}
\title{Get metadata for a specific dataset on LA City Open Data}
\usage{
lacity_metadata()
}
\arguments{
\item{dataset_id}{Socrata dataset identifier (e.g. "2nrs-mtv8")}
}
\value{tibble: id, name, description, category, type, updated_at,}
\description{Get metadata for a specific dataset on LA City Open Data}
